Added new execution extensions for schema
